Refactor import statements to replace 'shared' with 'types'
